requirements.txt
Flask==3.1.0
praw==7.8.1
python-dotenv==1.0.1


